correct TaskStar

+++ Seminar05/Task02/Program.cs
@@ -4,21 +4,21 @@
 //[-4, -6, 89, 6] -> 0
 
 
-int[] MonArray(int size, int min, int max)
+double[] MonArray(double size, double min, double max)
 {
-    int[] res = new int[size];
-    for (int i =0; i < size; i++)
+    double[] res = new double[size];
+    for (double i =0; i < size; i++)
         {
             res[i] = new Random().Next(min, max +1);
         }
     return res;
 }
 
-void ENNumbers(int[] array)
+void ENNumbers(double[] array)
 {
-    int res = 0;
+    double res = 0;
 
-    for (int i = 0; i < array.Length; i++)
+    for (double i = 0; i < array.Length; i++)
     {
         if (i % 2 != 0) res = res + array[i];
     }
@@ -26,12 +26,17 @@ void ENNumbers(int[] array)
 }
 
 Console.Write("Введите количество элементов в массиве: ");
-int number = int.Parse(Console.ReadLine());
+double number = double.Parse(Console.ReadLine());
+//double number = Convert.ToInt32(Console.ReadLine());
 Console.Write("Введите значение начала интервала: ");
-int beginline = int.Parse(Console.ReadLine());
+double beginline = double.Parse(Console.ReadLine());
+//double beginline = Convert.ToInt32(Console.ReadLine());
 Console.Write("Введите значение окончания интервала: ");
-int endline = int.Parse(Console.ReadLine());
+double endline = double.Parse(Console.ReadLine());
+//double endline = Convert.ToInt32(Console.ReadLine());
 
-int[] EvenArray = MonArray(number, beginline, endline);
+//Convert.ToInt32
+
+double[] EvenArray = MonArray(number, beginline, endline);
 Console.WriteLine(String.Join(" ", EvenArray));
 ENNumbers(EvenArray);
